Beginning with the final fixes:
All the screens are linked. Also the file that was lost (the recipe page that luka made) is
back. Arrow back buttons are also linked, except for the arrow back button on the recipe
guide page. What's left to be done is fixing the weird not happening click event on that
arrow back button that is not working, and add some transition between the pages, and if
there is time we should clean up our script.js file because there is too much of the same
code. Let's try that tomorrow.